feat: add Movie endpoints for user

+++ src/MoviePlatform.Domain/Common/PagedList.cs
@@ -5,12 +5,12 @@ public sealed class PagedList<T>
     public List<T> Items { get; }
     public int Page { get; }
     public int PageSize { get; }
-    public int TotalCount { get; }
+    public long TotalCount { get; }
     public int TotalPages => (int) Math.Ceiling((double) TotalCount / PageSize);
     public bool HasNextPage => Page < TotalPages;
     public bool HasPreviousPage => Page > 1;
 
-	private PagedList(List<T> items, int page, int pageSize, int totalCount)
+	private PagedList(List<T> items, int page, int pageSize, long totalCount)
     {
         Items = items;
         Page = page;
@@ -18,7 +18,7 @@ public sealed class PagedList<T>
         TotalCount = totalCount;
     }
 
-    public static PagedList<T> Create(List<T> items, int page, int pageSize, int totalCount)
+    public static PagedList<T> Create(List<T> items, int page, int pageSize, long totalCount)
     {
         return new(items, page, pageSize, totalCount);
     }
